requirements.txt
# pip install -r requirements.txt

selenium
webdriver-manager
pytest
pytest-html
pytest-xdist
allure-pytest
webcolors
pynput

Flask==2.3.3
python-dotenv==1.0.1
choco install ngrok
pip install psutil
pip install openai
openai>=1.0
pandas
selenium

